// Client/src/utils/types.ts
export interface Message {
  id: string;
  body: string;
  conversationId: string;
  senderId: string;
  createdAt: string;
  updatedAt: string;
  sender: Sender;
  fileId?: string | null;
  fileUrl?: string | null;
}

export interface Sender {
  id: string;
  joinedAt: string;
  userId: string;
  conversationId: string;
  role: string;
  user: User;
}
// ...
export interface MessageCardProps {
  message: Message;
  passRef: boolean;
  messagesEndRef: React.RefObject<HTMLDivElement>;
}